requirements.txt
django==4.0.6
djangorestframework==3.13.1
django-cors-headers==3.13.0
django-environ==0.9.0
django-redis==5.2.0
django-storages==1.12.3
celery==5.2.7
django-ckeditor==6.4.2
psycopg2==2.9.3
channels==3.0.5
argon2-cffi==21.3.0
boto3==1.24.44
gunicorn==20.1.0
whitenoise==6.2.0
requests==2.28.1
Pillow==9.2.0
mutagen==1.45.1
pytz==2022.1

pyJWT==2.1.0
djoser==2.1.0